backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
python-multipart==0.0.9
pillow==10.2.0
numpy==1.26.4
scikit-learn==1.4.0
sqlalchemy==2.0.27
aiosqlite==0.19.0
tensorflow-cpu>=2.15.0
opencv-python-headless>=4.9.0
